% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisFunctions.R
\name{pointwise.integrate}
\alias{pointwise.integrate}
\title{Calculate Area under the curve when the curve is defined by connect-the-dots}
\usage{
pointwise.integrate(data)
}
\arguments{
\item{data}{a data.frame with columns t and y}
}
\description{
Calculate Area under the curve when the curve is defined by connect-the-dots
}
\examples{
data <- data.frame( t = 1:10, y = rnorm(10))
pointwise.integrate(data)
}
